% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday Problem Probability}
\usage{
birthday(x)
}
\arguments{
\item{x}{a quantitative vector}
}
\value{
The probability that at least two people have the same birthday within a sample of x people
}
\description{
Birthday Problem Probability
}
\examples{
birthday(20:24)

}
