% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firedam.R
\docType{data}
\name{firedam}
\alias{firedam}
\title{FIREDAM.csv data}
\format{
## `firedam`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{Distance}{Damage}
  \item{num}{num}
  \item{number}{number}
  ...
}
}
\source{
<C:/RPACKAGES/MATH4753F25zachary/data/firedam.rda>
}
\usage{
firedam
}
\description{
A subset of data from the FIREDAM.csv data set ...
}
\keyword{datasets}
