% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{myclt}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{A numeric vector}

\item{iter}{A number}
}
\value{
histogram of sample means as well as the vector of sample means
}
\description{
myclt
}
\examples{
myclt(n=20, iter=10000)
}
