% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{a number}

\item{sigma}{a number}

\item{a}{a number less than mu +/- 3*sigma}
}
\value{
a normal curve plot with shaded area -infinity to a
}
\description{
myncurve
}
\examples{
myncurve(mu=5, sigma=2, a=3)
}
