% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysample.R
\name{mysample}
\alias{mysample}
\title{mysample}
\usage{
mysample(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{a numeric}

\item{iter}{a numeric}

\item{time}{a numeric float value}
}
\value{
a set of bargraphs
}
\description{
mysample
}
\examples{
mysample(n=1000, iter=30, time=1)
}
