% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{N tickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{numeric number of seats}

\item{gamma}{probability of overbooking}

\item{p}{probability of a ticket holder attending}
}
\value{
discrete and continuous n value and corresponding graphs
}
\description{
N tickets
}
\examples{
ntickets(N=400, gamma=0.02, p=0.95)
}
