% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterHist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter plot with marginal histograms.}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{a quantitative vector}

\item{y}{a quantitative vector}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
A scatter plot with x vs y and marginal histograms
}
\description{
Scatter plot with marginal histograms.
}
\examples{
scatterhist(x = 1:10, y = 11:20)
}
